refactor: isolate sidebar filtering
